// snapdeal code/paymentsmerchantpanel/MerChantPanel/MerchantPanelApp/src/main/webapp/src/app/payment/refund/refund-action/refund-action-message.interface.ts
/**
 * Created by shruti.aggarwal on 14/4/16.
 */
export interface IRefundActionMessage{
  requiredMessage: string;
  patternMessage: string;
  amountGreater: string;
  nonZeroValue: string;
  successMessage: string;
  errorMessage: string;
  successError: string;
  submitSuccess: string;
  tryAgain: string;
}